# backend/tests/baseline/openapi.json
{
  "components": {
    "schemas": {
...
      "ConnectionRequestCreate": {
        "properties": {
          "intent": {
            "anyOf": [
              {
                "maxLength": 50,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intent"
          },
          "note": {
            "anyOf": [
              {
                "maxLength": 240,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "receiver_id": {
            "format": "uuid",
            "title": "Receiver Id",
            "type": "string"
          }
        },
        "required": [
          "receiver_id"
        ],
        "title": "ConnectionRequestCreate",
        "type": "object"
      },
      "ConnectionRequestEnriched": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "intent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intent"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "partner_profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProfilePublic"
              },
              {
                "type": "null"
              }
            ]
          },
          "receiver_id": {
            "format": "uuid",
            "title": "Receiver Id",
            "type": "string"
          },
          "sender_id": {
            "format": "uuid",
            "title": "Sender Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "sender_id",
          "receiver_id",
          "intent",
          "note",
          "status",
          "created_at"
        ],
        "title": "ConnectionRequestEnriched",
        "type": "object"
      },
      "ConnectionRequestRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "intent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intent"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "receiver_id": {
            "format": "uuid",
            "title": "Receiver Id",
            "type": "string"
          },
          "sender_id": {
            "format": "uuid",
            "title": "Sender Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "sender_id",
          "receiver_id",
          "intent",
          "note",
          "status",
          "created_at"
        ],
        "title": "ConnectionRequestRead",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MatchRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "partner": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProfilePublic"
              },
              {
                "type": "null"
              }
            ]
          },
          "user_a_id": {
            "format": "uuid",
            "title": "User A Id",
            "type": "string"
          },
          "user_b_id": {
            "format": "uuid",
            "title": "User B Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "user_a_id",
          "user_b_id",
          "created_at"
        ],
        "title": "MatchRead",
        "type": "object"
      },
...
      "ProfilePublic": {
        "properties": {
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "bio": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bio"
          },
          "campus": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Campus"
          },
          "class_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Class Year"
          },
          "country_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country State"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "interests": {
            "items": {
              "type": "string"
            },
            "title": "Interests",
            "type": "array"
          },
          "is_hidden": {
            "title": "Is Hidden",
            "type": "boolean"
          },
          "is_verified": {
            "title": "Is Verified",
            "type": "boolean"
          },
          "languages_learning": {
            "items": {
              "type": "string"
            },
            "title": "Languages Learning",
            "type": "array"
          },
          "languages_spoken": {
            "items": {
              "type": "string"
            },
            "title": "Languages Spoken",
            "type": "array"
          },
          "looking_for": {
            "items": {
              "type": "string"
            },
            "title": "Looking For",
            "type": "array"
          },
          "looking_for_codes": {
            "items": {
              "type": "string"
            },
            "title": "Looking For Codes",
            "type": "array"
          },
          "major": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Major"
          },
          "profile_completed": {
            "title": "Profile Completed",
            "type": "boolean"
          },
          "pronouns": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pronouns"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "user_id",
          "display_name",
          "pronouns",
          "major",
          "class_year",
          "country_state",
          "campus",
          "bio",
          "avatar_url",
          "is_hidden",
          "is_verified",
          "profile_completed",
          "interests",
          "languages_spoken",
          "languages_learning",
          "looking_for",
          "looking_for_codes"
        ],
        "title": "ProfilePublic",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "LC Connect API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/connections/incoming": {
      "get": {
        "operationId": "incoming_requests_api_v1_connections_incoming_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConnectionRequestEnriched"
                  },
                  "title": "Response Incoming Requests Api V1 Connections Incoming Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Incoming Requests",
        "tags": [
          "connections"
        ]
      }
    },
    "/api/v1/connections/matches": {
      "get": {
        "operationId": "list_matches_api_v1_connections_matches_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MatchRead"
                  },
                  "title": "Response List Matches Api V1 Connections Matches Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Matches",
        "tags": [
          "connections"
        ]
      }
    },
    "/api/v1/connections/outgoing": {
      "get": {
        "operationId": "outgoing_requests_api_v1_connections_outgoing_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConnectionRequestEnriched"
                  },
                  "title": "Response Outgoing Requests Api V1 Connections Outgoing Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Outgoing Requests",
        "tags": [
          "connections"
        ]
      }
    },
    "/api/v1/connections/request": {
      "post": {
        "operationId": "send_connection_request_api_v1_connections_request_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionRequestCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionRequestRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Send Connection Request",
        "tags": [
          "connections"
        ]
      }
    },
    "/api/v1/connections/{request_id}/accept": {
      "post": {
        "operationId": "accept_request_api_v1_connections__request_id__accept_post",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Accept Request",
        "tags": [
          "connections"
        ]
      }
    },
    "/api/v1/connections/{request_id}/decline": {
      "post": {
        "operationId": "decline_request_api_v1_connections__request_id__decline_post",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionRequestRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Decline Request",
        "tags": [
          "connections"
        ]
      }
    },
...
  }
}